# docs/openapi.yaml
...
openapi: 3.1.0
info:
  title: HomeStock
  description: '

    A self-hosted home inventory system.


    Every response uses the same envelope: `{"data": ..., "error": null}`.

    On failure `data` is null and `error` holds a code and a message.


    The server runs every AI model on the CPU. There is no GPU. The AI routes

    return a job instead of a result when inference takes longer than

    `HS_AI_INLINE_TIMEOUT` seconds.

    '
  version: 0.1.0
paths:
...
  /api/receipts/upload:
    post:
      tags:
      - Receipts
      summary: Upload a receipt image and start the OCR.
      description: The row is created at once. The OCR runs on the CPU and may take a while, so ocr_parsed_json
        is null until it finishes. Poll the returned AI job, or fetch the receipt again.
      operationId: upload_receipt_api_receipts_upload_post
      requestBody:
        content:
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/Body_upload_receipt_api_receipts_upload_post'
        required: true
      responses:
        '201':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Envelope_ReceiptDetail_'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
      security:
      - HTTPBearer: []
  /api/receipts:
    get:
      tags:
      - Receipts
      summary: List receipts.
      operationId: list_receipts_api_receipts_get
      security:
      - HTTPBearer: []
      parameters:
      - name: vendor
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Vendor
      - name: page
        in: query
        required: false
        schema:
          type: integer
          minimum: 1
          default: 1
          title: Page
      - name: per_page
        in: query
        required: false
        schema:
          type: integer
          maximum: 200
          minimum: 1
          default: 50
          title: Per Page
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Envelope_Page_ReceiptRead__'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/receipts/{receipt_id}:
    get:
      tags:
      - Receipts
      summary: Return one receipt with its OCR output and its lines.
      operationId: get_receipt_api_receipts__receipt_id__get
      security:
      - HTTPBearer: []
      parameters:
      - name: receipt_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Receipt Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Envelope_ReceiptDetail_'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    put:
      tags:
      - Receipts
      summary: Correct the data that the OCR read.
      operationId: update_receipt_api_receipts__receipt_id__put
      security:
      - HTTPBearer: []
      parameters:
      - name: receipt_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Receipt Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ReceiptUpdate'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Envelope_ReceiptDetail_'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    delete:
      tags:
      - Receipts
      summary: Delete a receipt.
      operationId: delete_receipt_api_receipts__receipt_id__delete
      security:
      - HTTPBearer: []
      parameters:
      - name: receipt_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Receipt Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Envelope_Message_'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/receipts/{receipt_id}/link/{item_id}:
    post:
      tags:
      - Receipts
      summary: Link a receipt line to an inventory item.
      operationId: link_receipt_item_api_receipts__receipt_id__link__item_id__post
      security:
      - HTTPBearer: []
      parameters:
      - name: receipt_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Receipt Id
      - name: item_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Item Id
      - name: line_id
        in: query
        required: false
        schema:
          anyOf:
          - type: string
            format: uuid
          - type: 'null'
          description: The receipt line to link. Omit to link the whole receipt.
          title: Line Id
        description: The receipt line to link. Omit to link the whole receipt.
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Envelope_ReceiptDetail_'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
components:
  schemas:
...
    Body_upload_receipt_api_receipts_upload_post:
      properties:
        file:
          type: string
          format: binary
          title: File
          description: One image of a receipt.
      type: object
      required:
      - file
      title: Body_upload_receipt_api_receipts_upload_post
...
    Envelope_Message_:
      properties:
        data:
          anyOf:
          - $ref: '#/components/schemas/Message'
          - type: 'null'
        error:
          anyOf:
          - $ref: '#/components/schemas/ErrorDetail'
          - type: 'null'
      type: object
      title: Envelope[Message]
...
    Envelope_Page_ReceiptRead__:
      properties:
        data:
          anyOf:
          - $ref: '#/components/schemas/Page_ReceiptRead_'
          - type: 'null'
        error:
          anyOf:
          - $ref: '#/components/schemas/ErrorDetail'
          - type: 'null'
      type: object
      title: Envelope[Page[ReceiptRead]]
...
    Envelope_ReceiptDetail_:
      properties:
        data:
          anyOf:
          - $ref: '#/components/schemas/ReceiptDetail'
          - type: 'null'
        error:
          anyOf:
          - $ref: '#/components/schemas/ErrorDetail'
          - type: 'null'
      type: object
      title: Envelope[ReceiptDetail]
...
    ErrorDetail:
      properties:
        code:
          type: string
          title: Code
          description: A stable machine readable code.
        message:
          type: string
          title: Message
          description: A message for a person to read.
        details:
          anyOf:
          - type: object
          - type: 'null'
          title: Details
          description: Extra context, such as field errors.
      type: object
      required:
      - code
      - message
      title: ErrorDetail
      description: The error half of the response envelope.
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
...
    Message:
      properties:
        message:
          type: string
          title: Message
      type: object
      required:
      - message
      title: Message
      description: A plain acknowledgement, for routes that return no object.
...
    Page_ReceiptRead_:
      properties:
        items:
          items:
            $ref: '#/components/schemas/ReceiptRead'
          type: array
          title: Items
        total:
          type: integer
          title: Total
          description: Total rows that match the filter.
        page:
          type: integer
          title: Page
          description: The current page number. The first page is 1.
        per_page:
          type: integer
          title: Per Page
        pages:
          type: integer
          title: Pages
          description: Total number of pages.
      type: object
      required:
      - items
      - total
      - page
      - per_page
      - pages
      title: Page[ReceiptRead]
...
    ReceiptDetail:
      properties:
        id:
          type: string
          format: uuid
          title: Id
        user_id:
          type: string
          format: uuid
          title: User Id
        file_path:
          type: string
          title: File Path
        thumbnail_path:
          anyOf:
          - type: string
          - type: 'null'
          title: Thumbnail Path
        vendor:
          anyOf:
          - type: string
          - type: 'null'
          title: Vendor
        purchase_date:
          anyOf:
          - type: string
            format: date
          - type: 'null'
          title: Purchase Date
        total_amount:
          anyOf:
          - type: string
          - type: 'null'
          title: Total Amount
        currency:
          type: string
          title: Currency
        version:
          type: integer
          title: Version
        created_at:
          type: string
          format: date-time
          title: Created At
        updated_at:
          type: string
          format: date-time
          title: Updated At
        ocr_raw_text:
          anyOf:
          - type: string
          - type: 'null'
          title: Ocr Raw Text
        ocr_parsed_json:
          anyOf:
          - type: object
          - type: 'null'
          title: Ocr Parsed Json
        lines:
          items:
            $ref: '#/components/schemas/ReceiptLineRead'
          type: array
          title: Lines
      type: object
      required:
      - id
      - user_id
      - file_path
      - thumbnail_path
      - vendor
      - purchase_date
      - total_amount
      - currency
      - version
      - created_at
      - updated_at
      title: ReceiptDetail
      description: One receipt with its OCR output and its lines.
    ReceiptLineRead:
      properties:
        id:
          type: string
          format: uuid
          title: Id
        receipt_id:
          type: string
          format: uuid
          title: Receipt Id
        item_id:
          anyOf:
          - type: string
            format: uuid
          - type: 'null'
          title: Item Id
        line_text:
          type: string
          title: Line Text
        line_amount:
          anyOf:
          - type: string
          - type: 'null'
          title: Line Amount
        created_at:
          type: string
          format: date-time
          title: Created At
      type: object
      required:
      - id
      - receipt_id
      - item_id
      - line_text
      - line_amount
      - created_at
      title: ReceiptLineRead
      description: One line on a receipt.
    ReceiptRead:
      properties:
        id:
          type: string
          format: uuid
          title: Id
        user_id:
          type: string
          format: uuid
          title: User Id
        file_path:
          type: string
          title: File Path
        thumbnail_path:
          anyOf:
          - type: string
          - type: 'null'
          title: Thumbnail Path
        vendor:
          anyOf:
          - type: string
          - type: 'null'
          title: Vendor
        purchase_date:
          anyOf:
          - type: string
            format: date
          - type: 'null'
          title: Purchase Date
        total_amount:
          anyOf:
          - type: string
          - type: 'null'
          title: Total Amount
        currency:
          type: string
          title: Currency
        version:
          type: integer
          title: Version
        created_at:
          type: string
          format: date-time
          title: Created At
        updated_at:
          type: string
          format: date-time
          title: Updated At
      type: object
      required:
      - id
      - user_id
      - file_path
      - thumbnail_path
      - vendor
      - purchase_date
      - total_amount
      - currency
      - version
      - created_at
      - updated_at
      title: ReceiptRead
      description: A receipt in a list.
    ReceiptUpdate:
      properties:
        vendor:
          anyOf:
          - type: string
            maxLength: 200
          - type: 'null'
          title: Vendor
        purchase_date:
          anyOf:
          - type: string
            format: date
          - type: 'null'
          title: Purchase Date
        total_amount:
          anyOf:
          - type: number
            minimum: 0.0
          - type: string
          - type: 'null'
          title: Total Amount
        currency:
          anyOf:
          - type: string
            maxLength: 3
            minLength: 3
          - type: 'null'
          title: Currency
      type: object
      title: ReceiptUpdate
      description: Correct the data read from a receipt.
...
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          type: array
          title: Location
        msg:
          type: string
          title: Message
        type:
          type: string
          title: Error Type
      type: object
      required:
      - loc
      - msg
      - type
      title: ValidationError
  securitySchemes:
    HTTPBearer:
      type: http
      description: A JWT access token.
      scheme: bearer